+++ src/main/java/nl/b3p/jdbc/util/converter/GeometryJdbcConverterFactory.java
@@ -16,7 +16,7 @@
  * DEALINGS IN THE SOFTWARE.
  *
  */
-package nl.b3p.loader.jdbc;
+package nl.b3p.jdbc.util.converter;
 
 import oracle.jdbc.OracleConnection;
 import org.apache.commons.dbutils.QueryRunner;
