+++ src/test/java/nl/b3p/jdbc/util/converter/RoundTripIntegrationTest.java
@@ -16,9 +16,8 @@
  * DEALINGS IN THE SOFTWARE.
  *
  */
-package nl.b3p.loader.jdbc;
+package nl.b3p.jdbc.util.converter;
 
-import nl.b3p.AbstractDatabaseIntegrationTest;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.junit.jupiter.api.BeforeEach;
